requirements.txt
Django==3.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-extensions==2.2.8
django-ipware==2.1.0
django-markdownify==0.8.0
django-phonenumber-field==4.0.0
django-su==0.9.0
django-tables2==2.2.1
geoip2==3.0.0
openpyxl==3.0.3
phonenumbers==8.11.2
requests==2.22.0
requests-oauthlib==1.0.0
rules==2.2
tablib==1.1.0
